// cgocore-ui/src/app/features/models/shipments/views/full/AirWaybillFullView.ts
import { AirWaybillForm } from 'src/app/features/models/shipments/forms/AirWaybillForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirWaybillFullView {

id: number;
form: AirWaybillForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}